Refactor training UI states
