+++ apps/src/ui/docs/src/screens/training-config.md
@@ -25,10 +25,10 @@ scale 1.6
 
 ```plantuml
 @startuml
-[*] --> Training
+[*] --> TrainingIdle
 
-Training --> TrainingConfigOpen : Select Evolution
-TrainingConfigOpen --> Training : Close
-TrainingConfigOpen --> Training : Start
+TrainingIdle --> TrainingConfigOpen : Select Evolution
+TrainingConfigOpen --> TrainingIdle : Close
+TrainingConfigOpen --> TrainingActive : Start
 @enduml
 ```
