+++ apps/src/ui/state-machine/states/Paused.cpp
@@ -2,7 +2,6 @@
 #include "core/LoggingChannels.h"
 #include "core/network/WebSocketService.h"
 #include "server/api/SimStop.h"
-#include "ui/RemoteInputDevice.h"
 #include "ui/state-machine/StateMachine.h"
 #include "ui/ui_builders/LVGLBuilder.h"
 
@@ -124,56 +123,6 @@ void Paused::onQuitClicked(lv_event_t* e)
     UiApi::Exit::Cwc cwc;
     cwc.callback = [](auto&&) {};
     sm->queueEvent(cwc);
-}
-
-State::Any Paused::onEvent(const UiApi::Exit::Cwc& cwc, StateMachine& /*sm*/)
-{
-    LOG_INFO(State, "Exit command received, shutting down");
-
-    cwc.sendResponse(UiApi::Exit::Response::okay(std::monostate{}));
-
-    return Shutdown{};
-}
-
-State::Any Paused::onEvent(const UiApi::MouseDown::Cwc& cwc, StateMachine& sm)
-{
-    LOG_DEBUG(State, "Mouse down at ({}, {})", cwc.command.pixelX, cwc.command.pixelY);
-
-    // Update remote input device state (enables LVGL widget interaction even when paused).
-    if (sm.getRemoteInputDevice()) {
-        sm.getRemoteInputDevice()->updatePosition(cwc.command.pixelX, cwc.command.pixelY);
-        sm.getRemoteInputDevice()->updatePressed(true);
-    }
-
-    cwc.sendResponse(UiApi::MouseDown::Response::okay(std::monostate{}));
-    return Paused{ std::move(worldData) };
-}
-
-State::Any Paused::onEvent(const UiApi::MouseMove::Cwc& cwc, StateMachine& sm)
-{
-    LOG_DEBUG(State, "Mouse move at ({}, {})", cwc.command.pixelX, cwc.command.pixelY);
-
-    // Update remote input device position (enables LVGL widget interaction even when paused).
-    if (sm.getRemoteInputDevice()) {
-        sm.getRemoteInputDevice()->updatePosition(cwc.command.pixelX, cwc.command.pixelY);
-    }
-
-    cwc.sendResponse(UiApi::MouseMove::Response::okay(std::monostate{}));
-    return Paused{ std::move(worldData) };
-}
-
-State::Any Paused::onEvent(const UiApi::MouseUp::Cwc& cwc, StateMachine& sm)
-{
-    LOG_DEBUG(State, "Mouse up at ({}, {})", cwc.command.pixelX, cwc.command.pixelY);
-
-    // Update remote input device state (enables LVGL widget interaction even when paused).
-    if (sm.getRemoteInputDevice()) {
-        sm.getRemoteInputDevice()->updatePosition(cwc.command.pixelX, cwc.command.pixelY);
-        sm.getRemoteInputDevice()->updatePressed(false);
-    }
-
-    cwc.sendResponse(UiApi::MouseUp::Response::okay(std::monostate{}));
-    return Paused{ std::move(worldData) };
 }
 
 State::Any Paused::onEvent(const UiApi::SimRun::Cwc& cwc, StateMachine& /*sm*/)
