+++ apps/src/ui/state-machine/states/Paused.h
@@ -30,10 +30,6 @@ struct Paused {
     void onEnter(StateMachine& sm);
     void onExit(StateMachine& sm);
 
-    Any onEvent(const UiApi::Exit::Cwc& cwc, StateMachine& sm);
-    Any onEvent(const UiApi::MouseDown::Cwc& cwc, StateMachine& sm);
-    Any onEvent(const UiApi::MouseMove::Cwc& cwc, StateMachine& sm);
-    Any onEvent(const UiApi::MouseUp::Cwc& cwc, StateMachine& sm);
     Any onEvent(const UiApi::SimRun::Cwc& cwc, StateMachine& sm);
     Any onEvent(const UiApi::SimStop::Cwc& cwc, StateMachine& sm);
     Any onEvent(const ServerDisconnectedEvent& evt, StateMachine& sm);
