+++ apps/src/ui/state-machine/states/StartMenu.h
@@ -30,12 +30,8 @@ struct StartMenu {
     Any onEvent(const StartMenuIdleTimeoutEvent& evt, StateMachine& sm);
     Any onEvent(const TrainButtonClickedEvent& evt, StateMachine& sm);
     Any onEvent(const NextFractalClickedEvent& evt, StateMachine& sm);
-    Any onEvent(const UiApi::Exit::Cwc& cwc, StateMachine& sm);
     Any onEvent(const UiApi::SimRun::Cwc& cwc, StateMachine& sm);
     Any onEvent(const UiApi::TrainingStart::Cwc& cwc, StateMachine& sm);
-    Any onEvent(const UiApi::MouseDown::Cwc& cwc, StateMachine& sm);
-    Any onEvent(const UiApi::MouseMove::Cwc& cwc, StateMachine& sm);
-    Any onEvent(const UiApi::MouseUp::Cwc& cwc, StateMachine& sm);
 
     // Update background animations (fractal).
     void updateAnimations();
