+++ apps/src/ui/state-machine/states/Synth.cpp
@@ -1,7 +1,6 @@
 #include "State.h"
 #include "core/Assert.h"
 #include "core/LoggingChannels.h"
-#include "ui/RemoteInputDevice.h"
 #include "ui/UiComponentManager.h"
 #include "ui/state-machine/StateMachine.h"
 #include <string>
@@ -139,13 +138,6 @@ State::Any Synth::onEvent(const ServerDisconnectedEvent& evt, StateMachine& sm)
     return Disconnected{};
 }
 
-State::Any Synth::onEvent(const UiApi::Exit::Cwc& cwc, StateMachine& /*sm*/)
-{
-    LOG_INFO(State, "Exit command received, shutting down");
-    cwc.sendResponse(UiApi::Exit::Response::okay(std::monostate{}));
-    return Shutdown{};
-}
-
 State::Any Synth::onEvent(const UiApi::SimStop::Cwc& cwc, StateMachine& /*sm*/)
 {
     LOG_INFO(State, "SimStop command received, returning to StartMenu");
@@ -173,38 +165,6 @@ State::Any Synth::onEvent(const UiApi::SynthKeyPress::Cwc& cwc, StateMachine& /*
         .is_black = cwc.command.is_black,
     };
     cwc.sendResponse(UiApi::SynthKeyPress::Response::okay(response));
-    return std::move(*this);
-}
-
-State::Any Synth::onEvent(const UiApi::MouseDown::Cwc& cwc, StateMachine& sm)
-{
-    if (sm.getRemoteInputDevice()) {
-        sm.getRemoteInputDevice()->updatePosition(cwc.command.pixelX, cwc.command.pixelY);
-        sm.getRemoteInputDevice()->updatePressed(true);
-    }
-
-    cwc.sendResponse(UiApi::MouseDown::Response::okay({}));
-    return std::move(*this);
-}
-
-State::Any Synth::onEvent(const UiApi::MouseMove::Cwc& cwc, StateMachine& sm)
-{
-    if (sm.getRemoteInputDevice()) {
-        sm.getRemoteInputDevice()->updatePosition(cwc.command.pixelX, cwc.command.pixelY);
-    }
-
-    cwc.sendResponse(UiApi::MouseMove::Response::okay({}));
-    return std::move(*this);
-}
-
-State::Any Synth::onEvent(const UiApi::MouseUp::Cwc& cwc, StateMachine& sm)
-{
-    if (sm.getRemoteInputDevice()) {
-        sm.getRemoteInputDevice()->updatePosition(cwc.command.pixelX, cwc.command.pixelY);
-        sm.getRemoteInputDevice()->updatePressed(false);
-    }
-
-    cwc.sendResponse(UiApi::MouseUp::Response::okay({}));
     return std::move(*this);
 }
