+++ apps/src/ui/state-machine/states/TrainingIdle.cpp
@@ -1,0 +1,576 @@
+#include "TrainingIdle.h"
+#include "Disconnected.h"
+#include "SimRunning.h"
+#include "StartMenu.h"
+#include "State.h"
+#include "TrainingActive.h"
+#include "core/Assert.h"
+#include "core/LoggingChannels.h"
+#include "core/ScenarioConfig.h"
+#include "core/network/WebSocketService.h"
+#include "server/api/EvolutionStart.h"
+#include "server/api/GenomeGet.h"
+#include "server/api/SeedAdd.h"
+#include "server/api/SimRun.h"
+#include "ui/TrainingIdleView.h"
+#include "ui/UiComponentManager.h"
+#include "ui/state-machine/StateMachine.h"
+#include <algorithm>
+#include <lvgl.h>
+#include <utility>
+
+namespace DirtSim {
+namespace Ui {
+namespace State {
+
+TrainingIdle::TrainingIdle(TrainingSpec lastTrainingSpec, bool hasTrainingSpec)
+    : lastTrainingSpec_(std::move(lastTrainingSpec)), hasTrainingSpec_(hasTrainingSpec)
+{}
+
+TrainingIdle::~TrainingIdle() = default;
+
+void TrainingIdle::onEnter(StateMachine& sm)
+{
+    LOG_INFO(State, "Entering Training idle state (waiting for start command)");
+
+    auto* uiManager = sm.getUiComponentManager();
+    if (!uiManager) {
+        LOG_ERROR(State, "No UiComponentManager available");
+        return;
+    }
+
+    DirtSim::Network::WebSocketServiceInterface* wsService = nullptr;
+    if (sm.hasWebSocketService()) {
+        wsService = &sm.getWebSocketService();
+    }
+
+    view_ = std::make_unique<TrainingIdleView>(uiManager, sm, wsService, sm.getUserSettings());
+
+    if (auto* panel = uiManager->getExpandablePanel()) {
+        panel->clearContent();
+        panel->hide();
+        panel->resetWidth();
+    }
+
+    IconRail* iconRail = uiManager->getIconRail();
+    DIRTSIM_ASSERT(iconRail, "IconRail must exist");
+    if (lv_obj_t* railContainer = iconRail->getContainer()) {
+        lv_obj_clear_flag(railContainer, LV_OBJ_FLAG_HIDDEN);
+        lv_obj_clear_flag(railContainer, LV_OBJ_FLAG_IGNORE_LAYOUT);
+    }
+    iconRail->setLayout(RailLayout::SingleColumn);
+    iconRail->setVisibleIcons(
+        { IconId::DUCK, IconId::EVOLUTION, IconId::GENOME_BROWSER, IconId::TRAINING_RESULTS });
+    iconRail->deselectAll();
+
+    if (view_) {
+        view_->setEvolutionStarted(false);
+        view_->clearPanelContent();
+        view_->updateIconRailOffset();
+    }
+}
+
+void TrainingIdle::onExit(StateMachine& sm)
+{
+    LOG_INFO(State, "Exiting Training idle state");
+
+    view_.reset();
+
+    // Clear panel content after view cleanup.
+    if (auto* uiManager = sm.getUiComponentManager()) {
+        if (auto* panel = uiManager->getExpandablePanel()) {
+            panel->clearContent();
+            panel->hide();
+        }
+    }
+}
+
+void TrainingIdle::updateAnimations()
+{
+    if (view_) {
+        view_->updateAnimations();
+    }
+}
+
+bool TrainingIdle::isTrainingResultModalVisible() const
+{
+    return view_ && view_->isTrainingResultModalVisible();
+}
+
+State::Any TrainingIdle::onEvent(
+    const EvolutionProgressReceivedEvent& /*evt*/, StateMachine& /*sm*/)
+{
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(
+    const TrainingBestSnapshotReceivedEvent& /*evt*/, StateMachine& /*sm*/)
+{
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const IconSelectedEvent& evt, StateMachine& sm)
+{
+    LOG_INFO(
+        State,
+        "Icon selection changed: {} -> {}",
+        static_cast<int>(evt.previousId),
+        static_cast<int>(evt.selectedId));
+
+    if (!view_) {
+        return std::move(*this);
+    }
+
+    if (evt.selectedId == IconId::DUCK) {
+        LOG_INFO(State, "Start menu icon selected, returning to start menu");
+        if (auto* iconRail = sm.getUiComponentManager()->getIconRail()) {
+            iconRail->deselectAll();
+        }
+        return StartMenu{};
+    }
+
+    // Closing panel (deselected icon).
+    if (evt.selectedId == IconId::NONE) {
+        view_->clearPanelContent();
+        return std::move(*this);
+    }
+
+    view_->clearPanelContent();
+
+    switch (evt.selectedId) {
+        case IconId::EVOLUTION:
+            view_->createTrainingConfigPanel();
+            break;
+
+        case IconId::GENOME_BROWSER:
+            view_->createGenomeBrowserPanel();
+            break;
+
+        case IconId::TRAINING_RESULTS:
+            view_->createTrainingResultBrowserPanel();
+            break;
+
+        case IconId::TREE:
+        case IconId::MUSIC:
+        case IconId::NETWORK:
+        case IconId::DUCK:
+        case IconId::CORE:
+        case IconId::PHYSICS:
+        case IconId::PLAY:
+        case IconId::SCENARIO:
+        case IconId::NONE:
+            DIRTSIM_ASSERT(false, "Unexpected icon selection in Training idle");
+            break;
+    }
+
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const RailAutoShrinkRequestEvent& /*evt*/, StateMachine& sm)
+{
+    LOG_INFO(State, "Auto-shrink requested, minimizing IconRail");
+
+    if (auto* iconRail = sm.getUiComponentManager()->getIconRail()) {
+        iconRail->setMode(RailMode::Minimized);
+    }
+
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const ServerDisconnectedEvent& evt, StateMachine& sm)
+{
+    LOG_WARN(State, "Server disconnected during training (reason: {})", evt.reason);
+    LOG_INFO(State, "Transitioning to Disconnected");
+
+    if (!sm.queueReconnectToLastServer()) {
+        LOG_WARN(State, "No previous server address available for reconnect");
+    }
+
+    return Disconnected{};
+}
+
+State::Any TrainingIdle::onEvent(const StartEvolutionButtonClickedEvent& evt, StateMachine& sm)
+{
+    LOG_INFO(
+        State,
+        "Starting evolution: population={}, generations={}, scenario={}, organism_type={}",
+        evt.evolution.populationSize,
+        evt.evolution.maxGenerations,
+        toString(evt.training.scenarioId),
+        static_cast<int>(evt.training.organismType));
+
+    if (!sm.hasWebSocketService()) {
+        LOG_ERROR(State, "No WebSocketService available");
+        return std::move(*this);
+    }
+
+    auto& wsService = sm.getWebSocketService();
+    if (!wsService.isConnected()) {
+        LOG_WARN(State, "Not connected to server, cannot start evolution");
+        return std::move(*this);
+    }
+
+    Api::EvolutionStart::Command cmd;
+    cmd.evolution = evt.evolution;
+    cmd.mutation = evt.mutation;
+    cmd.scenarioId = evt.training.scenarioId;
+    cmd.organismType = evt.training.organismType;
+    cmd.population = evt.training.population;
+
+    const auto result =
+        wsService.sendCommandAndGetResponse<Api::EvolutionStart::OkayType>(cmd, 5000);
+    if (result.isError()) {
+        LOG_ERROR(State, "Failed to send EvolutionStart: {}", result.errorValue());
+        return std::move(*this);
+    }
+    if (result.value().isError()) {
+        LOG_ERROR(State, "Server EvolutionStart error: {}", result.value().errorValue().message);
+        return std::move(*this);
+    }
+
+    LOG_INFO(State, "Evolution started on server");
+    lastTrainingSpec_ = evt.training;
+    hasTrainingSpec_ = true;
+
+    return TrainingActive{ lastTrainingSpec_, hasTrainingSpec_ };
+}
+
+State::Any TrainingIdle::onEvent(const UiApi::TrainingStart::Cwc& cwc, StateMachine& sm)
+{
+    StartEvolutionButtonClickedEvent evt{
+        .evolution = cwc.command.evolution,
+        .mutation = cwc.command.mutation,
+        .training = cwc.command.training,
+    };
+    auto nextState = onEvent(evt, sm);
+    cwc.sendResponse(UiApi::TrainingStart::Response::okay({ .queued = true }));
+    return nextState;
+}
+
+State::Any TrainingIdle::onEvent(const UiApi::TrainingResultSave::Cwc& cwc, StateMachine& /*sm*/)
+{
+    cwc.sendResponse(
+        UiApi::TrainingResultSave::Response::error(ApiError("Training result modal not visible")));
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const UiApi::GenomeBrowserOpen::Cwc& cwc, StateMachine& sm)
+{
+    using Response = UiApi::GenomeBrowserOpen::Response;
+
+    if (!view_) {
+        cwc.sendResponse(Response::error(ApiError("Training view not available")));
+        return std::move(*this);
+    }
+
+    auto* uiManager = sm.getUiComponentManager();
+    if (!uiManager) {
+        cwc.sendResponse(Response::error(ApiError("UiComponentManager not available")));
+        return std::move(*this);
+    }
+
+    view_->clearPanelContent();
+    view_->createGenomeBrowserPanel();
+
+    if (auto* iconRail = uiManager->getIconRail()) {
+        iconRail->selectIcon(IconId::GENOME_BROWSER);
+    }
+
+    cwc.sendResponse(Response::okay({ .opened = true }));
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const UiApi::GenomeDetailOpen::Cwc& cwc, StateMachine& /*sm*/)
+{
+    using Response = UiApi::GenomeDetailOpen::Response;
+
+    if (!view_) {
+        cwc.sendResponse(Response::error(ApiError("Training view not available")));
+        return std::move(*this);
+    }
+
+    Result<GenomeId, std::string> result;
+    if (cwc.command.id.has_value()) {
+        result = view_->openGenomeDetailById(cwc.command.id.value());
+    }
+    else {
+        result = view_->openGenomeDetailByIndex(cwc.command.index);
+    }
+    if (result.isError()) {
+        cwc.sendResponse(Response::error(ApiError(result.errorValue())));
+        return std::move(*this);
+    }
+
+    UiApi::GenomeDetailOpen::Okay response{
+        .opened = true,
+        .id = result.value(),
+    };
+    cwc.sendResponse(Response::okay(std::move(response)));
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const UiApi::GenomeDetailLoad::Cwc& cwc, StateMachine& /*sm*/)
+{
+    using Response = UiApi::GenomeDetailLoad::Response;
+
+    if (!view_) {
+        cwc.sendResponse(Response::error(ApiError("Training view not available")));
+        return std::move(*this);
+    }
+
+    auto result = view_->loadGenomeDetail(cwc.command.id);
+    if (result.isError()) {
+        cwc.sendResponse(Response::error(ApiError(result.errorValue())));
+        return std::move(*this);
+    }
+
+    cwc.sendResponse(Response::okay({ .queued = true }));
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(
+    const UiApi::TrainingConfigShowEvolution::Cwc& cwc, StateMachine& /*sm*/)
+{
+    using Response = UiApi::TrainingConfigShowEvolution::Response;
+
+    if (!view_) {
+        cwc.sendResponse(Response::error(ApiError("Training view not available")));
+        return std::move(*this);
+    }
+
+    auto result = view_->showTrainingConfigView(TrainingIdleView::TrainingConfigView::Evolution);
+    if (result.isError()) {
+        cwc.sendResponse(Response::error(ApiError(result.errorValue())));
+        return std::move(*this);
+    }
+
+    cwc.sendResponse(Response::okay({ .opened = true }));
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const UiApi::TrainingQuit::Cwc& cwc, StateMachine& sm)
+{
+    auto nextState = onEvent(QuitTrainingClickedEvent{}, sm);
+    cwc.sendResponse(UiApi::TrainingQuit::Response::okay({ .queued = true }));
+    return nextState;
+}
+
+State::Any TrainingIdle::onEvent(const UiApi::TrainingResultDiscard::Cwc& cwc, StateMachine& /*sm*/)
+{
+    cwc.sendResponse(
+        UiApi::TrainingResultDiscard::Response::error(
+            ApiError("Training result modal not visible")));
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const TrainingStreamConfigChangedEvent& evt, StateMachine& sm)
+{
+    auto& settings = sm.getUserSettings();
+    settings.streamIntervalMs = std::max(0, evt.intervalMs);
+
+    if (view_) {
+        view_->setStreamIntervalMs(settings.streamIntervalMs);
+    }
+
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const StopTrainingClickedEvent& /*evt*/, StateMachine& /*sm*/)
+{
+    LOG_INFO(State, "Stop button clicked while idle, returning to start menu");
+    return StartMenu{};
+}
+
+State::Any TrainingIdle::onEvent(const QuitTrainingClickedEvent& /*evt*/, StateMachine& /*sm*/)
+{
+    LOG_INFO(State, "Quit button clicked, returning to start menu");
+    return StartMenu{};
+}
+
+State::Any TrainingIdle::onEvent(const GenomeLoadClickedEvent& evt, StateMachine& sm)
+{
+    LOG_INFO(State, "Genome load requested (genome_id={})", evt.genomeId.toShortString());
+
+    if (!sm.hasWebSocketService()) {
+        LOG_WARN(State, "Genome load ignored: no WebSocketService");
+        return std::move(*this);
+    }
+
+    auto& wsService = sm.getWebSocketService();
+    if (!wsService.isConnected()) {
+        LOG_WARN(State, "Genome load ignored: not connected to server");
+        return std::move(*this);
+    }
+
+    Api::GenomeGet::Command getCmd{ .id = evt.genomeId };
+    auto getResult = wsService.sendCommandAndGetResponse<Api::GenomeGet::Okay>(getCmd, 5000);
+    if (getResult.isError()) {
+        LOG_ERROR(State, "GenomeGet failed: {}", getResult.errorValue());
+        return std::move(*this);
+    }
+    if (getResult.value().isError()) {
+        LOG_ERROR(State, "GenomeGet error: {}", getResult.value().errorValue().message);
+        return std::move(*this);
+    }
+
+    const auto& response = getResult.value().value();
+    if (!response.found) {
+        LOG_WARN(State, "Genome load ignored: genome not found");
+        return std::move(*this);
+    }
+    if (!response.metadata.organismType.has_value()) {
+        LOG_WARN(State, "Genome load ignored: missing organism type");
+        return std::move(*this);
+    }
+    if (response.metadata.organismType.value() != OrganismType::TREE) {
+        LOG_WARN(State, "Genome load only supported for tree organisms");
+        return std::move(*this);
+    }
+
+    if (evt.genomeId.isNil()) {
+        LOG_WARN(State, "Genome load ignored: genome_id is nil");
+        return std::move(*this);
+    }
+
+    lv_disp_t* disp = lv_disp_get_default();
+    int16_t dispWidth = static_cast<int16_t>(lv_disp_get_hor_res(disp));
+    int16_t dispHeight = static_cast<int16_t>(lv_disp_get_ver_res(disp));
+    Vector2s containerSize{ static_cast<int16_t>(dispWidth - IconRail::MINIMIZED_RAIL_WIDTH),
+                            dispHeight };
+
+    Api::SimRun::Command simRunCmd{ .timestep = 0.016,
+                                    .max_steps = -1,
+                                    .max_frame_ms = 16,
+                                    .scenario_id = evt.scenarioId,
+                                    .start_paused = false,
+                                    .container_size = containerSize };
+
+    auto simResult = wsService.sendCommandAndGetResponse<Api::SimRun::Okay>(simRunCmd, 2000);
+    if (simResult.isError() || simResult.value().isError()) {
+        LOG_ERROR(State, "SimRun failed");
+        return std::move(*this);
+    }
+
+    constexpr int targetCellSize = 16;
+    const int worldWidth = std::max(10, static_cast<int>(containerSize.x) / targetCellSize);
+    const int worldHeight = std::max(10, static_cast<int>(containerSize.y) / targetCellSize);
+    const int centerX = worldWidth / 2;
+    const int centerY = worldHeight / 2;
+
+    Api::SeedAdd::Command seedCmd;
+    seedCmd.x = centerX;
+    seedCmd.y = centerY;
+    seedCmd.genome_id = evt.genomeId.toString();
+
+    auto seedResult = wsService.sendCommandAndGetResponse<Api::SeedAdd::OkayType>(seedCmd, 2000);
+    if (seedResult.isError() || seedResult.value().isError()) {
+        LOG_ERROR(State, "SeedAdd failed");
+    }
+
+    LOG_INFO(State, "Transitioning to SimRunning with genome");
+    return SimRunning{};
+}
+
+State::Any TrainingIdle::onEvent(
+    const OpenTrainingGenomeBrowserClickedEvent& /*evt*/, StateMachine& sm)
+{
+    if (!view_) {
+        LOG_WARN(State, "Training view not available for genome browser");
+        return std::move(*this);
+    }
+
+    auto* uiManager = sm.getUiComponentManager();
+    if (!uiManager) {
+        LOG_WARN(State, "UiComponentManager not available for genome browser");
+        return std::move(*this);
+    }
+
+    view_->clearPanelContent();
+    view_->createGenomeBrowserPanel();
+
+    if (auto* iconRail = uiManager->getIconRail()) {
+        iconRail->selectIcon(IconId::GENOME_BROWSER);
+    }
+
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const GenomeAddToTrainingClickedEvent& evt, StateMachine& /*sm*/)
+{
+    if (!view_) {
+        LOG_WARN(State, "Training view not available for genome add");
+        return std::move(*this);
+    }
+
+    view_->addGenomeToTraining(evt.genomeId, evt.scenarioId);
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const UiUpdateEvent& /*evt*/, StateMachine& /*sm*/)
+{
+    return std::move(*this);
+}
+
+State::Any TrainingIdle::onEvent(const ViewBestButtonClickedEvent& evt, StateMachine& sm)
+{
+    LOG_INFO(State, "View Best clicked, genome_id={}", evt.genomeId.toShortString());
+
+    if (!hasTrainingSpec_) {
+        LOG_WARN(State, "View Best ignored: no training spec available");
+        return std::move(*this);
+    }
+    if (lastTrainingSpec_.organismType != OrganismType::TREE) {
+        LOG_WARN(State, "View Best only supported for tree training");
+        return std::move(*this);
+    }
+    if (evt.genomeId.isNil()) {
+        LOG_WARN(State, "View Best ignored: genome_id is nil");
+        return std::move(*this);
+    }
+
+    DIRTSIM_ASSERT(sm.hasWebSocketService(), "WebSocketService must exist");
+    auto& wsService = sm.getWebSocketService();
+    DIRTSIM_ASSERT(wsService.isConnected(), "Must be connected");
+
+    lv_disp_t* disp = lv_disp_get_default();
+    int16_t dispWidth = static_cast<int16_t>(lv_disp_get_hor_res(disp));
+    int16_t dispHeight = static_cast<int16_t>(lv_disp_get_ver_res(disp));
+    Vector2s containerSize{ static_cast<int16_t>(dispWidth - IconRail::MINIMIZED_RAIL_WIDTH),
+                            dispHeight };
+
+    Api::SimRun::Command simRunCmd{ .timestep = 0.016,
+                                    .max_steps = -1,
+                                    .max_frame_ms = 16,
+                                    .scenario_id = lastTrainingSpec_.scenarioId,
+                                    .start_paused = false,
+                                    .container_size = containerSize };
+
+    auto simResult = wsService.sendCommandAndGetResponse<Api::SimRun::Okay>(simRunCmd, 2000);
+    if (simResult.isError() || simResult.value().isError()) {
+        LOG_ERROR(State, "SimRun failed");
+        return std::move(*this);
+    }
+
+    constexpr int targetCellSize = 16;
+    const int worldWidth = std::max(10, static_cast<int>(containerSize.x) / targetCellSize);
+    const int worldHeight = std::max(10, static_cast<int>(containerSize.y) / targetCellSize);
+    const int centerX = worldWidth / 2;
+    const int centerY = worldHeight / 2;
+
+    Api::SeedAdd::Command seedCmd;
+    seedCmd.x = centerX;
+    seedCmd.y = centerY;
+    seedCmd.genome_id = evt.genomeId.toString();
+
+    auto seedResult = wsService.sendCommandAndGetResponse<Api::SeedAdd::OkayType>(seedCmd, 2000);
+    if (seedResult.isError() || seedResult.value().isError()) {
+        LOG_ERROR(State, "SeedAdd failed");
+    }
+
+    LOG_INFO(State, "Transitioning to SimRunning with best genome");
+    return SimRunning{};
+}
+
+} // namespace State
+} // namespace Ui
+} // namespace DirtSim
